+++ client/QuestionsAnswers/Answers.jsx
@@ -1,13 +1,6 @@
 import React from 'react';
 import moment from 'moment';
-const paragraphStyle = {
-  display: 'inline',
-  fontSize: 12,
-  color: 'grey',
-  marginLeft: 30,
-  marginBottom: 10,
-  marginTop: 10
-};
+import {paragraphStyle} from './QAstyle.jsx';
 
 function Answers({answer}) {
   return (
